Move main as __init__ to avoid name conflicts
